+++ src/datasources/results/query-handlers/query-results/QueryResultsDataSource.ts
@@ -35,9 +35,6 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
   }
 
   async runQuery(query: QueryResults, options: DataQueryRequest): Promise<DataFrameDTO> {
-    await this.getPartNumbers();
-    await this.loadWorkspaces();
-
     if (query.queryBy) {
       query.queryBy = transformComputedFieldsQuery(
         this.templateSrv.replace(query.queryBy, options.scopedVars),