+++ src/datasources/work-orders/constants/QueryEditor.constants.ts
@@ -1,0 +1,7 @@
+export const tooltips = {
+    queryBy: 'This optional field specifies the query filters.',
+    outputType: 'This field specifies the output type to fetch work order properties or total count',
+    properties: 'This field specifies the properties to use in the query.',
+    orderBy: 'This field specifies the query order of the work orders.',
+    descending: 'This toggle returns the work orders query in descending order.',
+};